backend/requirements_loose.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
alphashape
amqp
annotated-doc
annotated-types
anyio
APScheduler
asyncpg
attrs
backoff
beautifulsoup4
billiard
brotli
cachetools
celery
certifi
cffi
chardet
charset-normalizer
chromadb
click
click-didyoumean
click-log
click-plugins
click-repl
clickhouse-connect
Crawl4AI
cryptography
cssselect
diskcache
distro
duckdb
exceptiongroup
faiss-cpu
fake-http-header
fake-useragent
fastapi
fasttext-wheel
fastuuid
feedparser
filelock
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h2
hf-xet
hnswlib
hpack
httpcore
httplib2
httptools
httpx
huggingface-hub
humanize
hyperframe
idna
importlib_metadata
iniconfig
Jinja2
jiter
joblib
jsonschema
jsonschema-specifications
kombu
lark
litellm
llama_cpp_python
lxml
lz4
markdown-it-py
MarkupSafe
mdurl
mpmath
multidict
networkx
nltk
numpy
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
patchright
pillow
playwright
pluggy
posthog
prompt_toolkit
propcache
proto-plus
protobuf
psutil
pyasn1
pyasn1_modules
pybase64
pybind11
pycparser
pydantic
pydantic-settings
pydantic_core
pyee
Pygments
pyOpenSSL
pyparsing
pytest
pytest-asyncio
python-dateutil
python-dotenv
pytz
PyYAML
rank-bm25
redis
referencing
regex
requests
rich
rpds-py
rsa
rtree
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
sgmllib3k
shapely
six
sniffio
snowballstemmer
soupsieve
SQLAlchemy
starlette
sympy
tenacity
tf-playwright-stealth
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
trimesh
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
uvicorn
vine
watchfiles
wcwidth
websockets
wheel
xxhash
yarl
zipp
zstandard
